+++ src/main/java/com/example/whathis/auth/dto/response/TokenResponse.java
@@ -1,4 +1,4 @@
-package com.example.whathis.user.dto.response;
+package com.example.whathis.auth.dto.response;
 
 import lombok.Builder;
 import lombok.Getter;
